Fix build

+++ submodules/MtProtoKit/MTRsa.m
@@ -295,7 +295,7 @@ static NSData *base64_decode(NSString *str) {
                                &outlen
                                );
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
             ret = nil;
             break;
         }else{
@@ -351,7 +351,7 @@ static NSData *base64_decode(NSString *str) {
                                &outlen
                                );
         if (status != 0) {
-            NSLog(@"SecKeyEncrypt fail. Error Code: %d", status);
+            NSLog(@"SecKeyEncrypt fail. Error Code: %d", (int)status);
             ret = nil;
             break;
         }else{
